recreated all api basiclly from scratch

+++ app/db/user.py
@@ -1,7 +1,7 @@
 from typing import Optional
 from app.models.user import OrmUserData, UserData
 from sqlalchemy.orm import Session
-from app.db.models import OneTimeKey, User
+from app.db.models import User
 from .crypto import set_one_time_keys
 
 
@@ -23,10 +23,7 @@ def add_user(db: Session, user: UserData) -> bool:
     db.add(item)
     db.commit()
 
-    if (
-        set_one_time_keys(db, user.login, {i: v for i, v in enumerate(keys)})
-        is False
-    ):
+    if set_one_time_keys(db, user.login, {i: v for i, v in enumerate(keys)}) is False:
         return False
     return True
 
@@ -35,14 +32,18 @@ def delete_user(db: Session, login: str) -> bool:
     if get_user_data(db, login) is None:
         return False
 
+    if db.query is None:
+        return False
+
+    if db.query(User) is None:
+        return False
+
     db.query(User).filter(User.login == login).delete()
     db.commit()
     return True
 
 
-def get_user_data(
-    db: Session, login: str, get_id: bool = False
-) -> Optional[UserData]:
+def get_user_data(db: Session, login: str, get_id: bool = False) -> Optional[UserData]:
     result = db.query(User).filter(User.login == login).first()
 
     if result is None:
